tests/mse_mp4/1342913/t.html

<!DOCTYPE html>
<html><head>
<meta http-equiv="content-type" content="text/html; charset=windows-1252">
  <title>MSE: resume from waiting even after format change occurred</title>
  <script type="text/javascript" src="../common.js"></script>
  <script type="text/javascript" src="../mediasource.js"></script>
</head>
<body>
<pre id="test"><script class="testbody" type="text/javascript">

runWithMSE(function(ms, el) {
  el.controls = true;
  once(ms, 'sourceopen').then(function() {
    ok(true, "Receive a sourceopen event");
    var sb = ms.addSourceBuffer("video/mp4");
    fetchAndLoad(sb, '../bipbop/bipbop_480_624kbps-video', ['init'], '.mp4')
    .then(fetchAndLoad.bind(null, sb, '../bipbop/bipbop_480_624kbps-video', range(1, 3), '.m4s'))
    .then(function() {
      el.play();
      // let seek to the last audio frame.
      el.currentTime = 1.596677;
      return Promise.all([once(el, 'seeked'), once(el, 'stalled')]);
    })
    .then(function() {
      info("seek completed");
      return fetchAndLoad(sb, '../bipbop/bipbop', ['init'], '.mp4');
    })
    .then(fetchAndLoad.bind(null, sb, '../bipbop/bipbop', range(1, 4), '.m4s'))
    .then(function() {
      ms.endOfStream();
      return once(el, 'ended');
    }).then(function() {
      SimpleTest.finish();
    });
  });
});

</script>
</pre>
</body>
</html>
